requirements.txt
anthropic
boto3 >= 1.34.131, < 1.35.0
bs4
chromedriver-py >= 128.0.6613.137
cohere >= 5.9.2
docopt >= 0.6.2
faiss-cpu >= 1.8.0
google-api-python-client >= 2.145.0
pdfplumber >= 0.11.4
python-dotenv >= 1.0.1
langchain >= 0.3.0
langchain-aws >= 0.2.0
langchain-fireworks
langchain_core >= 0.3.0
langchain-cohere
langchain_community
langchain_experimental
langchain_openai
langchain-ollama
langchain_groq
langchain-google-genai
langchain-together
langsmith
lxml >= 5.3.0
lxml_html_clean >= 0.2.2
schema
streamlit >= 1.38.0
selenium
tiktoken
transformers >= 4.44.2
rich >= 13.8.1
trafilatura >= 1.12.2
watchdog >= 2.1.5, < 5.0.0
spacy >= 3.6.1, < 4.0.0